requirements.txt
python-dotenv
py-cord
gpt4all